Cambiando el tipo de dato de las cuentas del status de las tareas

+++ src/main/java/com/metaphorce/assessment_final/dto/TaskStatusCount.java
@@ -6,5 +6,5 @@ public interface TaskStatusCount {
 
     Status getStatus();
 
-    long getCount();
+    int getCount();
 }
